Refactor VisNetwork into partial classes
